+++ src/Types/StrictusEnum.php
@@ -4,9 +4,9 @@ declare(strict_types=1);
 
 namespace Strictus\Types;
 
+use Strictus\Concerns\StrictusTyping;
 use Strictus\Exceptions\StrictusTypeException;
 use Strictus\Interfaces\StrictusTypeInterface;
-use Strictus\Traits\StrictusTyping;
 
 /**
  * @internal
